+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Your changes/patches go here.
 
+- [CHORE: Remove GPL licensed dependency Colorize and replace it with Rainbow]
+
 # v1.3.0 / 2023-06-16 [(commits)](https://github.com/fastruby/next_rails/compare/v1.2.4...v1.3.0)
 
 - [FEATURE: Add NextRails.next? for application usage (e.g. Rails shims)](https://github.com/fastruby/next_rails/pull/97)