+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-# main [(unreleased)](https://github.com/fastruby/next_rails/compare/v1.3.0...main)
+# main [(unreleased)](https://github.com/fastruby/next_rails/compare/v1.4.0...main)
 
 * Your changes/patches go here.
-- [CHORE: Use next_rails namespace on spec tests.](https://github.com/fastruby/next_rails/pull/117)
 
+# v1.4.0 / 2024-09-24 [(commits)](https://github.com/fastruby/next_rails/compare/v1.3.0...v1.4.0)
+
+- [CHORE: Use next_rails namespace on spec tests.](https://github.com/fastruby/next_rails/pull/117)
 - [CHORE: Remove 2.0.0, 2.1, 2.2 Ruby support](https://github.com/fastruby/next_rails/pull/126)
 - [CHORE: Update compatibility for Ruby versions to use Rainbow](https://github.com/fastruby/next_rails/pull/125)
 - [FEATURE: Support compatibility for Ruby versions](https://github.com/fastruby/next_rails/pull/116)